requirements.txt
pyserial
behave
selenium
Appium-Python-Client
allure-behave